Fix bug in searches routing
There was a bug with the regexp that performs the URL parameter matching
because it can't contain parenthesis, so we have to remove it in order
to avoid Angular to crash in some situations.

Also, I have added compatibility to search ':' characters without being
ignored by the parameter parsing code.